Break cd() into its own library
